fix: battling cypress' dashboard feature (#30441)
